Refactor CLI responsibilities into SRP modules
