+++ crates/uselesskey-cli/src/main.rs
@@ -1,11 +1,10 @@
 #![forbid(unsafe_code)]
 
 use std::fs;
-use std::io::{self, Read, Write};
 use std::path::{Path, PathBuf};
 
 use anyhow::{Context, Result, bail};
-use clap::{Parser, Subcommand, ValueEnum};
+use clap::Parser;
 use serde::{Deserialize, Serialize};
 use serde_json::json;
 use uselesskey_cli::{
@@ -14,251 +13,24 @@ use uselesskey_cli::{
     render_k8s_secret_yaml, render_vault_kv_json,
 };
 use uselesskey_core::Factory;
-use uselesskey_ecdsa::{EcdsaFactoryExt, EcdsaSpec};
-use uselesskey_ed25519::{Ed25519FactoryExt, Ed25519Spec};
-use uselesskey_hmac::{HmacFactoryExt, HmacSpec};
 use uselesskey_jwk::NegativeJwks;
 use uselesskey_rsa::{RsaFactoryExt, RsaSpec};
 use uselesskey_token::{NegativeToken, TokenFactoryExt, TokenSpec};
 use uselesskey_webhook::{WebhookFactoryExt, WebhookPayloadSpec};
-use uselesskey_x509::{ChainNegative, ChainSpec, X509Chain, X509FactoryExt, X509Spec};
+use uselesskey_x509::{ChainNegative, ChainSpec, X509Chain, X509FactoryExt};
 
-#[derive(Parser, Debug)]
-#[command(name = "uselesskey", about = "Deterministic fixture generation CLI")]
-struct Cli {
-    #[command(subcommand)]
-    command: Commands,
-}
+mod srp;
 
-#[derive(Subcommand, Debug)]
-enum Commands {
-    Generate(GenerateArgs),
-    Profiles(ProfilesArgs),
-    Profile(ProfileArgs),
-    Bundle(BundleArgs),
-    VerifyBundle(VerifyBundleArgs),
-    InspectBundle(InspectBundleArgs),
-    Export(ExportArgs),
-    Inspect(InspectArgs),
-    Materialize(MaterializeArgs),
-    Verify(VerifyArgs),
-}
-
-#[derive(clap::Args, Debug)]
-struct ProfilesArgs {
-    #[arg(long)]
-    explain: bool,
-}
-
-#[derive(clap::Args, Debug)]
-struct ProfileArgs {
-    profile: BundleProfile,
-    #[arg(long)]
-    explain: bool,
-}
-
-#[derive(clap::Args, Debug)]
-struct GenerateArgs {
-    kind: Kind,
-    #[arg(long)]
-    seed: String,
-    #[arg(long)]
-    label: String,
-    #[arg(long)]
-    format: Format,
-    #[arg(long)]
-    out: Option<PathBuf>,
-}
-
-#[derive(clap::Args, Debug)]
-struct BundleArgs {
-    #[arg(long, default_value = "uselesskey-bundle-seed")]
-    seed: String,
-    #[arg(long, default_value = "bundle")]
-    label: String,
-    #[arg(long, default_value = "jwk")]
-    format: Format,
-    #[arg(long, default_value = "scanner-safe")]
-    profile: BundleProfile,
-    #[arg(long)]
-    out: Option<PathBuf>,
-    #[arg(long)]
-    explain: bool,
-}
-
-#[derive(clap::Args, Debug)]
-struct VerifyBundleArgs {
-    #[arg(long = "bundle-dir", alias = "path")]
-    bundle_dir: PathBuf,
-}
-
-#[derive(clap::Args, Debug)]
-struct InspectBundleArgs {
-    #[arg(long = "bundle-dir", alias = "path")]
-    bundle_dir: PathBuf,
-    #[arg(long)]
-    out: Option<PathBuf>,
-}
-
-#[derive(clap::Args, Debug)]
-struct ExportArgs {
-    #[command(subcommand)]
-    target: ExportTarget,
-}
-
-#[derive(Subcommand, Debug)]
-enum ExportTarget {
-    K8s(ExportK8sArgs),
-    VaultKvJson(ExportVaultKvJsonArgs),
-}
-
-#[derive(clap::Args, Debug)]
-struct ExportK8sArgs {
-    #[arg(long = "bundle-dir", alias = "path")]
-    bundle_dir: PathBuf,
-    #[arg(long)]
-    name: String,
-    #[arg(long)]
-    namespace: Option<String>,
-    #[arg(long)]
-    out: Option<PathBuf>,
-}
-
-#[derive(clap::Args, Debug)]
-struct ExportVaultKvJsonArgs {
-    #[arg(long = "bundle-dir", alias = "path")]
-    bundle_dir: PathBuf,
-    #[arg(long)]
-    out: Option<PathBuf>,
-}
-
-#[derive(clap::Args, Debug)]
-struct InspectArgs {
-    #[arg(long)]
-    format: Format,
-    #[arg(long)]
-    input: Option<PathBuf>,
-    #[arg(long)]
-    out: Option<PathBuf>,
-}
-
-#[derive(clap::Args, Debug)]
-struct MaterializeArgs {
-    #[arg(long)]
-    manifest: PathBuf,
-    #[arg(long = "out-dir", alias = "out")]
-    out_dir: Option<PathBuf>,
-    #[arg(long)]
-    emit_rs: Option<PathBuf>,
-    #[arg(long, hide = true)]
-    check: bool,
-}
-
-#[derive(clap::Args, Debug)]
-struct VerifyArgs {
-    #[arg(long)]
-    manifest: PathBuf,
-    #[arg(long = "out-dir", alias = "out")]
-    out_dir: Option<PathBuf>,
-}
-
-#[derive(Clone, Copy, Debug, ValueEnum)]
-enum Kind {
-    Rsa,
-    Ecdsa,
-    Ed25519,
-    Hmac,
-    Token,
-    X509,
-    Jwk,
-    Jwks,
-}
-
-impl Kind {
-    const fn manifest_name(self) -> &'static str {
-        match self {
-            Self::Rsa => "rsa",
-            Self::Ecdsa => "ecdsa",
-            Self::Ed25519 => "ed25519",
-            Self::Hmac => "hmac",
-            Self::Token => "token",
-            Self::X509 => "x509",
-            Self::Jwk => "jwk",
-            Self::Jwks => "jwks",
-        }
-    }
-}
-
-#[derive(Clone, Copy, Debug, ValueEnum)]
-enum Format {
-    Pem,
-    Der,
-    Jwk,
-    Jwks,
-    #[value(name = "json-manifest")]
-    JsonManifest,
-    #[value(name = "bundle-dir")]
-    BundleDir,
-}
-
-impl Format {
-    const fn manifest_name(self) -> &'static str {
-        match self {
-            Self::Pem => "pem",
-            Self::Der => "der",
-            Self::Jwk => "jwk",
-            Self::Jwks => "jwks",
-            Self::JsonManifest => "json-manifest",
-            Self::BundleDir => "bundle-dir",
-        }
-    }
-}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq, ValueEnum)]
-enum BundleProfile {
-    ScannerSafe,
-    Oidc,
-    Tls,
-    Webhook,
-    Runtime,
-}
-
-impl BundleProfile {
-    const fn manifest_name(self) -> &'static str {
-        match self {
-            Self::ScannerSafe => "scanner-safe",
-            Self::Oidc => "oidc",
-            Self::Tls => "tls",
-            Self::Webhook => "webhook",
-            Self::Runtime => "runtime",
-        }
-    }
-
-    const fn output_dir_hint(self) -> &'static str {
-        match self {
-            Self::ScannerSafe => "target/uselesskey-bundle",
-            Self::Oidc => "target/uselesskey-oidc",
-            Self::Tls => "target/uselesskey-tls",
-            Self::Webhook => "target/uselesskey-webhook",
-            Self::Runtime => "target/uselesskey-runtime",
-        }
-    }
-}
-
-const DISCOVERABLE_PROFILES: [BundleProfile; 5] = [
-    BundleProfile::ScannerSafe,
-    BundleProfile::Tls,
-    BundleProfile::Oidc,
-    BundleProfile::Webhook,
-    BundleProfile::Runtime,
-];
-
-#[derive(Debug)]
-enum Artifact {
-    Text(String),
-    Binary(Vec<u8>),
-    Json(serde_json::Value),
-}
+use srp::artifact::{
+    Artifact, artifact_bytes, detect_kind, emit_artifact, format_extension, generate_artifact,
+    read_input, unsupported, write_artifact_to_path,
+};
+use srp::cli::{
+    BundleArgs, BundleProfile, Cli, Commands, ExportArgs, ExportK8sArgs, ExportTarget,
+    ExportVaultKvJsonArgs, Format, GenerateArgs, InspectArgs, InspectBundleArgs, Kind,
+    MaterializeArgs, ProfileArgs, ProfilesArgs, VerifyArgs, VerifyBundleArgs,
+};
+use srp::profiles::{render_profile_explanation, render_profile_summary, render_profiles};
 
 fn main() -> Result<()> {
     let cli = Cli::parse();
@@ -689,242 +461,6 @@ fn yes_no_unknown(value: Option<bool>) -> &'static str {
 
 fn count_or_unknown(value: Option<usize>) -> String {
     value.map_or_else(|| "unknown".to_string(), |count| count.to_string())
-}
-
-#[derive(Clone, Copy)]
-struct ProfileInfo {
-    profile: BundleProfile,
-    title: &'static str,
-    purpose: &'static str,
-    required_feature: &'static str,
-    scanner_posture: &'static str,
-    proof_command: &'static str,
-    claim: Option<&'static str>,
-    docs: &'static str,
-    generates: &'static [&'static str],
-    proves: &'static [&'static str],
-    not_proves: &'static [&'static str],
-}
-
-fn profile_info(profile: BundleProfile) -> ProfileInfo {
-    match profile {
-        BundleProfile::ScannerSafe => ProfileInfo {
-            profile,
-            title: "Scanner-safe baseline bundle",
-            purpose: "baseline scanner-safe fixtures, receipts, and export handoff metadata",
-            required_feature: "uselesskey-cli default features",
-            scanner_posture: "scanner-safe fixture material; generated exports still belong under target/",
-            proof_command: "cargo xtask claim-proof --claim scanner-safe-fixtures",
-            claim: Some("scanner-safe-fixtures"),
-            docs: "docs/how-to/generate-scanner-safe-k8s-secret.md",
-            generates: &[
-                "RSA/ECDSA/Ed25519/HMAC public fixture shapes",
-                "token and X.509 fixture shapes",
-                "manifest.json",
-                "receipts/materialization.json",
-                "receipts/audit-surface.json",
-            ],
-            proves: &[
-                "repo policy found no committed secret-shaped fixture blobs",
-                "the bundle has a manifest and audit receipts",
-                "scanner-safe badge drift checks still agree with policy",
-            ],
-            not_proves: &[
-                "every derived encoded export is safe to commit",
-                "production key management",
-                "scanner evasion",
-                "cryptographic assurance",
-            ],
-        },
-        BundleProfile::Tls => ProfileInfo {
-            profile,
-            title: "TLS contract pack",
-            purpose: "TLS chain and certificate rejection fixtures",
-            required_feature: "uselesskey-cli default features",
-            scanner_posture: "generated PEM payloads stay under target/; proof receipts are metadata",
-            proof_command: "cargo xtask claim-proof --claim tls-contract-pack",
-            claim: Some("tls-contract-pack"),
-            docs: "docs/how-to/test-tls-chain-validation.md",
-            generates: &[
-                "certs/valid-leaf.pem",
-                "certs/valid-chain.pem",
-                "expired, not-yet-valid, wrong-hostname, and untrusted-root negatives",
-                "evidence/tls-profile.md",
-                "manifest and receipts",
-            ],
-            proves: &[
-                "documented TLS fixture files are generated",
-                "positive and negative certificate paths are present",
-                "receipts and evidence docs are present",
-            ],
-            not_proves: &[
-                "production PKI",
-                "revocation, OCSP, certificate transparency, or mTLS",
-                "browser trust-store behavior",
-                "downstream verifier correctness",
-            ],
-        },
-        BundleProfile::Oidc => ProfileInfo {
-            profile,
-            title: "OIDC/JWKS contract pack",
-            purpose: "OIDC/JWKS validator fixtures and JWT-shaped negatives",
-            required_feature: "uselesskey-cli default features",
-            scanner_posture: "generated token/JWKS payloads stay under target/; proof receipts are metadata",
-            proof_command: "cargo xtask bundle-proof --profile oidc --out target/release-evidence/oidc",
-            claim: Some("oidc-jwks-contract-pack"),
-            docs: "docs/how-to/test-oidc-jwks-validation.md",
-            generates: &[
-                "valid JWKS and RS256 JWT-shaped fixtures",
-                "duplicate-kid and missing-kid JWKS negatives",
-                "alg-none and bad-audience token negatives",
-                "manifest and receipts",
-            ],
-            proves: &[
-                "deterministic JWKS and JWT-shaped fixtures are generated",
-                "documented validator negative inputs exist",
-                "receipts and evidence docs are present",
-            ],
-            not_proves: &[
-                "production signing-key custody",
-                "full OpenID provider behavior",
-                "issuer policy",
-                "downstream validator correctness",
-            ],
-        },
-        BundleProfile::Webhook => ProfileInfo {
-            profile,
-            title: "Webhook contract pack",
-            purpose: "HMAC webhook signature positives and negatives",
-            required_feature: "uselesskey-cli default features",
-            scanner_posture: "generated request payloads stay under target/; proof receipts are metadata",
-            proof_command: "cargo xtask claim-proof --claim webhook-contract-pack",
-            claim: Some("webhook-contract-pack"),
-            docs: "docs/how-to/test-webhook-signature-validation.md",
-            generates: &[
-                "requests/valid.json",
-                "tampered-body, wrong-secret, stale-timestamp, missing-signature, and malformed-signature negatives",
-                "evidence/webhook-profile.md",
-                "manifest and receipts",
-            ],
-            proves: &[
-                "deterministic HMAC verifier fixture behavior",
-                "valid signature acceptance and documented rejection classes",
-                "receipts and evidence docs are present",
-            ],
-            not_proves: &[
-                "provider compatibility",
-                "production secret management",
-                "replay protection completeness",
-                "delivery retries or transport security",
-                "downstream verifier correctness",
-            ],
-        },
-        BundleProfile::Runtime => ProfileInfo {
-            profile,
-            title: "Runtime bundle",
-            purpose: "general runtime fixture bundle for local experimentation",
-            required_feature: "uselesskey-cli default features",
-            scanner_posture: "may include runtime fixture material; keep generated payloads under target/",
-            proof_command: "uselesskey verify-bundle --path target/uselesskey-runtime",
-            claim: None,
-            docs: "README.md",
-            generates: &[
-                "general JWK/JWKS/X.509 fixture outputs",
-                "manifest.json",
-                "receipts/materialization.json",
-                "receipts/audit-surface.json",
-            ],
-            proves: &[
-                "the bundle can be regenerated and verified against its manifest",
-                "local output shape is internally consistent",
-            ],
-            not_proves: &[
-                "a public contract-pack claim",
-                "scanner-safe commit posture",
-                "production security behavior",
-            ],
-        },
-    }
-}
-
-fn render_profiles(explain: bool) -> String {
-    let mut out = String::new();
-    out.push_str("Available uselesskey profiles\n\n");
-    out.push_str("| Profile | Purpose | Proof |\n");
-    out.push_str("|---|---|---|\n");
-    for profile in DISCOVERABLE_PROFILES {
-        let info = profile_info(profile);
-        out.push_str(&format!(
-            "| `{}` | {} | `{}` |\n",
-            info.profile.manifest_name(),
-            info.purpose,
-            info.proof_command
-        ));
-    }
-    out.push_str("\nUse `uselesskey profile <name> --explain` for generated files, boundaries, and copyable commands.\n");
-
-    if explain {
-        out.push('\n');
-        for profile in DISCOVERABLE_PROFILES {
-            out.push_str(&render_profile_explanation(profile));
-            out.push('\n');
-        }
-    }
-
-    out
-}
-
-fn render_profile_summary(profile: BundleProfile) -> String {
-    let info = profile_info(profile);
-    format!(
-        concat!(
-            "Profile: {}\n",
-            "Title: {}\n",
-            "Purpose: {}\n",
-            "Generate: uselesskey bundle --profile {} --out {}\n",
-            "Verify: uselesskey verify-bundle --path {}\n",
-            "Proof: {}\n",
-            "Explain: uselesskey profile {} --explain\n",
-            "Bundle explain: uselesskey bundle --profile {} --explain\n",
-        ),
-        info.profile.manifest_name(),
-        info.title,
-        info.purpose,
-        info.profile.manifest_name(),
-        info.profile.output_dir_hint(),
-        info.profile.output_dir_hint(),
-        info.proof_command,
-        info.profile.manifest_name(),
-        info.profile.manifest_name(),
-    )
-}
-
-fn render_profile_explanation(profile: BundleProfile) -> String {
-    let info = profile_info(profile);
-    let mut out = render_profile_summary(profile);
-    out.push_str(&format!("Required feature: {}\n", info.required_feature));
-    out.push_str(&format!(
-        "Scanner/runtime posture: {}\n",
-        info.scanner_posture
-    ));
-    if let Some(claim) = info.claim {
-        out.push_str(&format!("Claim: {claim}\n"));
-    }
-    out.push_str(&format!("Docs: {}\n", info.docs));
-    push_list(&mut out, "\nGenerates", info.generates);
-    push_list(&mut out, "\nProves", info.proves);
-    push_list(&mut out, "\nDoes not prove", info.not_proves);
-    out
-}
-
-fn push_list(out: &mut String, title: &str, items: &[&str]) {
-    out.push_str(title);
-    out.push_str(":\n");
-    for item in items {
-        out.push_str("- ");
-        out.push_str(item);
-        out.push('\n');
-    }
 }
 
 #[cfg(test)]
@@ -1754,192 +1290,6 @@ fn generate_scanner_safe_bundle_artifact(
             }
         }
         _ => generate_artifact(fx, kind, label, format),
-    }
-}
-
-fn generate_artifact(fx: &Factory, kind: Kind, label: &str, format: Format) -> Result<Artifact> {
-    match kind {
-        Kind::Rsa => {
-            let kp = fx.rsa(label, RsaSpec::rs256());
-            match format {
-                Format::Pem => Ok(Artifact::Text(kp.private_key_pkcs8_pem().to_string())),
-                Format::Der => Ok(Artifact::Binary(kp.private_key_pkcs8_der().to_vec())),
-                Format::Jwk => Ok(Artifact::Json(kp.public_jwk_json())),
-                Format::Jwks => Ok(Artifact::Json(kp.public_jwks_json())),
-                _ => unsupported(kind, format),
-            }
-        }
-        Kind::Ecdsa => {
-            let kp = fx.ecdsa(label, EcdsaSpec::es256());
-            match format {
-                Format::Pem => Ok(Artifact::Text(kp.private_key_pkcs8_pem().to_string())),
-                Format::Der => Ok(Artifact::Binary(kp.private_key_pkcs8_der().to_vec())),
-                Format::Jwk => Ok(Artifact::Json(kp.public_jwk_json())),
-                Format::Jwks => Ok(Artifact::Json(kp.public_jwks_json())),
-                _ => unsupported(kind, format),
-            }
-        }
-        Kind::Ed25519 => {
-            let kp = fx.ed25519(label, Ed25519Spec::new());
-            match format {
-                Format::Pem => Ok(Artifact::Text(kp.private_key_pkcs8_pem().to_string())),
-                Format::Der => Ok(Artifact::Binary(kp.private_key_pkcs8_der().to_vec())),
-                Format::Jwk => Ok(Artifact::Json(kp.public_jwk_json())),
-                Format::Jwks => Ok(Artifact::Json(kp.public_jwks_json())),
-                _ => unsupported(kind, format),
-            }
-        }
-        Kind::Hmac => {
-            let sec = fx.hmac(label, HmacSpec::hs256());
-            match format {
-                Format::Der => Ok(Artifact::Binary(sec.secret_bytes().to_vec())),
-                Format::Jwk => Ok(Artifact::Json(sec.jwk().to_value())),
-                Format::Jwks => Ok(Artifact::Json(sec.jwks().to_value())),
-                _ => unsupported(kind, format),
-            }
-        }
-        Kind::Token => {
-            let token = fx.token(label, TokenSpec::api_key());
-            match format {
-                Format::Pem => Ok(Artifact::Text(token.value().to_string())),
-                Format::JsonManifest => Ok(Artifact::Json(
-                    json!({"kind":"token","label":label,"value":token.value()}),
-                )),
-                _ => unsupported(kind, format),
-            }
-        }
-        Kind::X509 => {
-            let cert = fx.x509_self_signed(label, X509Spec::self_signed(label));
-            match format {
-                Format::Pem => Ok(Artifact::Text(cert.cert_pem().to_string())),
-                Format::Der => Ok(Artifact::Binary(cert.cert_der().to_vec())),
-                _ => unsupported(kind, format),
-            }
-        }
-        Kind::Jwk => {
-            let kp = fx.rsa(label, RsaSpec::rs256());
-            if matches!(format, Format::Jwk) {
-                Ok(Artifact::Json(kp.public_jwk_json()))
-            } else {
-                unsupported(kind, format)
-            }
-        }
-        Kind::Jwks => {
-            let kp = fx.rsa(label, RsaSpec::rs256());
-            if matches!(format, Format::Jwks) {
-                Ok(Artifact::Json(kp.public_jwks_json()))
-            } else {
-                unsupported(kind, format)
-            }
-        }
-    }
-}
-
-fn unsupported(kind: Kind, format: Format) -> Result<Artifact> {
-    bail!("unsupported format {format:?} for kind {kind:?}")
-}
-
-fn emit_artifact(artifact: &Artifact, out: Option<&Path>) -> Result<()> {
-    if let Some(path) = out {
-        write_artifact_to_path(artifact, path)
-    } else {
-        write_artifact_to_stdout(artifact)
-    }
-}
-
-fn write_artifact_to_stdout(artifact: &Artifact) -> Result<()> {
-    let mut out = io::stdout().lock();
-    match artifact {
-        Artifact::Text(t) => out.write_all(t.as_bytes())?,
-        Artifact::Binary(b) => out.write_all(b)?,
-        Artifact::Json(v) => {
-            serde_json::to_writer_pretty(&mut out, v)?;
-            out.write_all(b"\n")?;
-        }
-    }
-    out.flush()?;
-    Ok(())
-}
-
-fn artifact_bytes(artifact: &Artifact) -> Result<Vec<u8>> {
-    match artifact {
-        Artifact::Text(t) => Ok(t.as_bytes().to_vec()),
-        Artifact::Binary(b) => Ok(b.clone()),
-        Artifact::Json(v) => Ok(serde_json::to_vec_pretty(v)?),
-    }
-}
-
-fn write_artifact_to_path(artifact: &Artifact, path: &Path) -> Result<()> {
-    if let Some(parent) = path.parent() {
-        fs::create_dir_all(parent)?;
-    }
-    fs::write(path, artifact_bytes(artifact)?)?;
-    Ok(())
-}
-
-fn read_input(path: Option<&Path>) -> Result<Vec<u8>> {
-    let mut buf = Vec::new();
-    match path {
-        Some(p) if p != Path::new("-") => {
-            buf = fs::read(p).with_context(|| format!("failed to read {}", p.display()))?
-        }
-        _ => {
-            io::stdin()
-                .lock()
-                .read_to_end(&mut buf)
-                .context("failed reading stdin")?;
-        }
-    }
-    Ok(buf)
-}
-
-fn format_extension(format: Format, artifact: &Artifact) -> &'static str {
-    match format {
-        Format::Pem => "pem",
-        Format::Der => "der",
-        Format::Jwk => "jwk.json",
-        Format::Jwks => "jwks.json",
-        Format::JsonManifest => "json",
-        Format::BundleDir => match artifact {
-            Artifact::Binary(_) => "bin",
-            Artifact::Json(_) => "json",
-            Artifact::Text(_) => "txt",
-        },
-    }
-}
-
-fn detect_kind(text: &str) -> &'static str {
-    if text.contains("BEGIN CERTIFICATE") {
-        "x509"
-    } else if text.contains("BEGIN PRIVATE KEY") {
-        "private_key"
-    } else {
-        detect_json_kind(text).unwrap_or("unknown")
-    }
-}
-
-fn detect_json_kind(text: &str) -> Option<&'static str> {
-    let trimmed = text.trim_start();
-    if !trimmed.starts_with('{') {
-        return None;
-    }
-
-    let value: serde_json::Value = serde_json::from_str(trimmed).ok()?;
-    let object = value.as_object()?;
-    if object
-        .get("keys")
-        .and_then(serde_json::Value::as_array)
-        .is_some()
-    {
-        Some("jwks")
-    } else if object
-        .get("kty")
-        .and_then(serde_json::Value::as_str)
-        .is_some()
-    {
-        Some("jwk")
-    } else {
-        None
     }
 }
