+++ crates/uselesskey-cli/src/srp/artifact.rs
@@ -1,0 +1,212 @@
+use std::fs;
+use std::io::{self, Read, Write};
+use std::path::Path;
+
+use anyhow::{Context, Result, bail};
+use serde_json::json;
+use uselesskey_core::Factory;
+use uselesskey_ecdsa::{EcdsaFactoryExt, EcdsaSpec};
+use uselesskey_ed25519::{Ed25519FactoryExt, Ed25519Spec};
+use uselesskey_hmac::{HmacFactoryExt, HmacSpec};
+use uselesskey_rsa::{RsaFactoryExt, RsaSpec};
+use uselesskey_token::{TokenFactoryExt, TokenSpec};
+use uselesskey_x509::{X509FactoryExt, X509Spec};
+
+use super::cli::{Format, Kind};
+
+pub(crate) enum Artifact {
+    Text(String),
+    Binary(Vec<u8>),
+    Json(serde_json::Value),
+}
+
+pub(crate) fn generate_artifact(
+    fx: &Factory,
+    kind: Kind,
+    label: &str,
+    format: Format,
+) -> Result<Artifact> {
+    match kind {
+        Kind::Rsa => {
+            let kp = fx.rsa(label, RsaSpec::rs256());
+            match format {
+                Format::Pem => Ok(Artifact::Text(kp.private_key_pkcs8_pem().to_string())),
+                Format::Der => Ok(Artifact::Binary(kp.private_key_pkcs8_der().to_vec())),
+                Format::Jwk => Ok(Artifact::Json(kp.public_jwk_json())),
+                Format::Jwks => Ok(Artifact::Json(kp.public_jwks_json())),
+                _ => unsupported(kind, format),
+            }
+        }
+        Kind::Ecdsa => {
+            let kp = fx.ecdsa(label, EcdsaSpec::es256());
+            match format {
+                Format::Pem => Ok(Artifact::Text(kp.private_key_pkcs8_pem().to_string())),
+                Format::Der => Ok(Artifact::Binary(kp.private_key_pkcs8_der().to_vec())),
+                Format::Jwk => Ok(Artifact::Json(kp.public_jwk_json())),
+                Format::Jwks => Ok(Artifact::Json(kp.public_jwks_json())),
+                _ => unsupported(kind, format),
+            }
+        }
+        Kind::Ed25519 => {
+            let kp = fx.ed25519(label, Ed25519Spec::new());
+            match format {
+                Format::Pem => Ok(Artifact::Text(kp.private_key_pkcs8_pem().to_string())),
+                Format::Der => Ok(Artifact::Binary(kp.private_key_pkcs8_der().to_vec())),
+                Format::Jwk => Ok(Artifact::Json(kp.public_jwk_json())),
+                Format::Jwks => Ok(Artifact::Json(kp.public_jwks_json())),
+                _ => unsupported(kind, format),
+            }
+        }
+        Kind::Hmac => {
+            let sec = fx.hmac(label, HmacSpec::hs256());
+            match format {
+                Format::Der => Ok(Artifact::Binary(sec.secret_bytes().to_vec())),
+                Format::Jwk => Ok(Artifact::Json(sec.jwk().to_value())),
+                Format::Jwks => Ok(Artifact::Json(sec.jwks().to_value())),
+                _ => unsupported(kind, format),
+            }
+        }
+        Kind::Token => {
+            let token = fx.token(label, TokenSpec::api_key());
+            match format {
+                Format::Pem => Ok(Artifact::Text(token.value().to_string())),
+                Format::JsonManifest => Ok(Artifact::Json(
+                    json!({"kind":"token","label":label,"value":token.value()}),
+                )),
+                _ => unsupported(kind, format),
+            }
+        }
+        Kind::X509 => {
+            let cert = fx.x509_self_signed(label, X509Spec::self_signed(label));
+            match format {
+                Format::Pem => Ok(Artifact::Text(cert.cert_pem().to_string())),
+                Format::Der => Ok(Artifact::Binary(cert.cert_der().to_vec())),
+                _ => unsupported(kind, format),
+            }
+        }
+        Kind::Jwk => {
+            let kp = fx.rsa(label, RsaSpec::rs256());
+            if matches!(format, Format::Jwk) {
+                Ok(Artifact::Json(kp.public_jwk_json()))
+            } else {
+                unsupported(kind, format)
+            }
+        }
+        Kind::Jwks => {
+            let kp = fx.rsa(label, RsaSpec::rs256());
+            if matches!(format, Format::Jwks) {
+                Ok(Artifact::Json(kp.public_jwks_json()))
+            } else {
+                unsupported(kind, format)
+            }
+        }
+    }
+}
+
+pub(crate) fn unsupported(kind: Kind, format: Format) -> Result<Artifact> {
+    bail!("unsupported format {format:?} for kind {kind:?}")
+}
+
+pub(crate) fn emit_artifact(artifact: &Artifact, out: Option<&Path>) -> Result<()> {
+    if let Some(path) = out {
+        write_artifact_to_path(artifact, path)
+    } else {
+        write_artifact_to_stdout(artifact)
+    }
+}
+
+fn write_artifact_to_stdout(artifact: &Artifact) -> Result<()> {
+    let mut out = io::stdout().lock();
+    match artifact {
+        Artifact::Text(t) => out.write_all(t.as_bytes())?,
+        Artifact::Binary(b) => out.write_all(b)?,
+        Artifact::Json(v) => {
+            serde_json::to_writer_pretty(&mut out, v)?;
+            out.write_all(b"\n")?;
+        }
+    }
+    out.flush()?;
+    Ok(())
+}
+
+pub(crate) fn artifact_bytes(artifact: &Artifact) -> Result<Vec<u8>> {
+    match artifact {
+        Artifact::Text(t) => Ok(t.as_bytes().to_vec()),
+        Artifact::Binary(b) => Ok(b.clone()),
+        Artifact::Json(v) => Ok(serde_json::to_vec_pretty(v)?),
+    }
+}
+
+pub(crate) fn write_artifact_to_path(artifact: &Artifact, path: &Path) -> Result<()> {
+    if let Some(parent) = path.parent() {
+        fs::create_dir_all(parent)?;
+    }
+    fs::write(path, artifact_bytes(artifact)?)?;
+    Ok(())
+}
+
+pub(crate) fn read_input(path: Option<&Path>) -> Result<Vec<u8>> {
+    let mut buf = Vec::new();
+    match path {
+        Some(p) if p != Path::new("-") => {
+            buf = fs::read(p).with_context(|| format!("failed to read {}", p.display()))?
+        }
+        _ => {
+            io::stdin()
+                .lock()
+                .read_to_end(&mut buf)
+                .context("failed reading stdin")?;
+        }
+    }
+    Ok(buf)
+}
+
+pub(crate) fn format_extension(format: Format, artifact: &Artifact) -> &'static str {
+    match format {
+        Format::Pem => "pem",
+        Format::Der => "der",
+        Format::Jwk => "jwk.json",
+        Format::Jwks => "jwks.json",
+        Format::JsonManifest => "json",
+        Format::BundleDir => match artifact {
+            Artifact::Binary(_) => "bin",
+            Artifact::Json(_) => "json",
+            Artifact::Text(_) => "txt",
+        },
+    }
+}
+
+pub(crate) fn detect_kind(text: &str) -> &'static str {
+    if text.contains("BEGIN CERTIFICATE") {
+        "x509"
+    } else if text.contains("BEGIN PRIVATE KEY") {
+        "private_key"
+    } else {
+        detect_json_kind(text).unwrap_or("unknown")
+    }
+}
+
+fn detect_json_kind(text: &str) -> Option<&'static str> {
+    let trimmed = text.trim_start();
+    if !trimmed.starts_with('{') {
+        return None;
+    }
+
+    let value: serde_json::Value = serde_json::from_str(trimmed).ok()?;
+    let object = value.as_object()?;
+    if object
+        .get("keys")
+        .and_then(serde_json::Value::as_array)
+        .is_some()
+    {
+        Some("jwks")
+    } else if object
+        .get("kty")
+        .and_then(serde_json::Value::as_str)
+        .is_some()
+    {
+        Some("jwk")
+    } else {
+        None
+    }
+}
