+++ crates/uselesskey-cli/src/srp/mod.rs
@@ -1,0 +1,3 @@
+pub(crate) mod artifact;
+pub(crate) mod cli;
+pub(crate) mod profiles;
